Improve parser resilience
A hostchar currently consists of a regex explicitly allowing certain
characters. This is not a very pleasing solution as networks may use
all kinds of characters in their host strings. Furthermore, the
whitelist approach serves no real purpose as the structure of a
message guarantees that the hostchars will be followed by a space.

The new solution is just simply allow any non-whitespace characters as
hostchars.